separeta streaming files
