+++ src/SerenityStar/Models/Streaming/StreamingAgentMessageStart.cs
@@ -1,0 +1,24 @@
+using System;
+
+namespace SerenityStar.Models.Streaming
+{
+    /// <summary>
+    /// Represents the start of a streaming session.
+    /// </summary>
+    public sealed class StreamingAgentMessageStart : StreamingAgentMessage
+    {
+        /// <summary>
+        /// The UTC time when the stream started.
+        /// </summary>
+        public DateTime StartTimeUtc { get; set; }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="StreamingAgentMessageStart"/> class.
+        /// </summary>
+        public StreamingAgentMessageStart()
+        {
+            Type = "start";
+            StartTimeUtc = DateTime.UtcNow;
+        }
+    }
+}
